rewrote the script to accomodate kinetic measurement
The first itearation of the script was made to deal with files produced
in point measurements. This version deals with multiple measurements
over a long period of time. Additionally, it also takes a "label" file,
which contains labels for each well. The labels are added to the
respective well entries.

Finally, the optparse package was integrated for more elegant control.